clusternetwork.js and clustersOfInterest.js linting bugfixes
