+++ .eslintrc
@@ -10,7 +10,8 @@
   },
   "env": {
     "browser": true,
-    "node": true
+    "node": true,
+    "es6": true
   },
   "rules": {
     "strict": "off",
@@ -43,6 +44,7 @@
   "globals": {
     "_": false,
     "d3": false,
-    "$": false
+    "$": false,
+    "__": false
   }
 }
