+++ .github/workflows/playwright.yml
@@ -16,10 +16,10 @@ jobs:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v4
-    - name: Use Node.js '19.x'
+    - name: Use Node.js '18.x'
       uses: actions/setup-node@v4
       with:
-        node-version: '19.x'
+        node-version: '18.x'
     - name: Install dependencies
       run: npm install -g yarn && yarn
     - name: Build the project
