+++ package.json
@@ -42,7 +42,7 @@
     "babel-preset-react-hmre": "^1.1.1",
     "concurrently": "^7.6.0",
     "css-loader": "^6.7.2",
-    "eslint": "^8.28.0",
+    "eslint": "^9.2.0",
     "exports-loader": "^4.0.0",
     "expose-loader": "^4.0.0",
     "express": "^4.18.2",
